src/backend/requirements.txt
ray[default]
google-cloud-bigquery
scikit-learn
google-cloud-storage
google-auth
db-dtypes
google-cloud-bigquery[pandas]
tensorflow
pandas